Separating out path variables into separate document
